test(types): expect suint length for dynamic shielded arrays
Update all tests involving dynamic shielded arrays to expect .length
to return suint256 instead of uint256, and to use cload/cstore instead
of sload/sstore for the length slot in inline assembly.

Add new semantic and syntax tests for comprehensive coverage of the
suint length behavior including type checking, cload/cstore assembly
verification, push/pop operations, and mapping interactions.

+++ test/libsolidity/semanticTests/array/copying/shielded_array_copy_sbool_storage.sol
@@ -12,7 +12,7 @@ contract C {
     }
 
     function getLengthB() public view returns (uint256) {
-        return b.length;
+        return uint256(b.length);
     }
 
     function getB(uint256 i) public view returns (bool) {
